refactor(player-portal): split api/types.ts into focused files
Replaces the 724-line monolithic types.ts with a types/ directory of
16 domain-focused files (primitives, stats, resources, defenses,
biography, movement, crafting, strikes, feats, actions, class, items,
spells, character, prepared) plus a barrel index.ts that re-exports
everything. Zero behavior change; all consumer imports resolve unchanged
via bundler directory resolution.
