+++ apps/player-portal/src/api/types/stats.ts
@@ -1,0 +1,76 @@
+import type { Modifier, ProficiencyRank } from './primitives';
+
+export interface SkillStatistic {
+  slug: string;
+  label: string;
+  value: number;
+  totalModifier: number;
+  dc: number;
+  breakdown: string;
+  modifiers: Modifier[];
+  rank: ProficiencyRank;
+  attribute: string;
+  armor?: boolean;
+  itemId?: string | null;
+  lore?: boolean;
+}
+
+export interface MartialProficiency {
+  rank: ProficiencyRank;
+  value: number;
+  breakdown: string;
+  visible?: boolean;
+  custom?: boolean;
+  label?: string;
+}
+
+export interface ClassDC {
+  slug: string;
+  label: string;
+  rank: ProficiencyRank;
+  attribute: string;
+  primary: boolean;
+  value: number;
+  totalModifier: number;
+  dc: number;
+  breakdown: string;
+  modifiers: Modifier[];
+}
+
+export interface SpellcastingProficiency {
+  rank: ProficiencyRank;
+}
+
+export interface Save {
+  slug: string;
+  label: string;
+  value: number;
+  totalModifier: number;
+  dc: number;
+  rank: ProficiencyRank;
+  attribute: string;
+  breakdown: string;
+  modifiers: Modifier[];
+}
+
+export interface Perception {
+  slug: string;
+  label: string;
+  value: number;
+  totalModifier: number;
+  dc: number;
+  rank: ProficiencyRank;
+  attribute: string;
+  breakdown: string;
+  modifiers: Modifier[];
+}
+
+export interface Initiative {
+  slug: string;
+  label: string;
+  value: number;
+  totalModifier: number;
+  breakdown: string;
+  statistic: string;
+  tiebreakPriority: number;
+}
